AGENT-600: Add e2e test to exercise agent-tui check and fix (#1524)
* AGENT-600: Add e2e test to exercise agent-tui check and fix

Adds a new configuration option AGENT_E2E_TEST_TUI_BAD_DNS to
enable a test scenario where the DNS is disabled on the hosts
by configuring the host to use an incorrect IP address for DNS.

This causes the agent-tui to reach the state where it shows the
release image cannot be pulled and waits for a user to fix it
the issue through the console.

The test then invokes a series of console key presses to simulate
a user fixing the issue through the console. The key presses
should fix the network configuration and allow the automated
flow to continue.

This test will allow us to catch the case where the agent-tui fails
to run but automated flow is still able to continue to success.
If the agent-tui fails to run now and if the automated flow is able
to still continue, it will not succeed because the broken DNS will
not allow the release image to be reached.

* Refactor bad_dns test scenario

Renamed AGENT_E2E_TEST_TUI_BAD_DNS to AGENT_TEST_CASES.
AGENT_TEST_CASES can contain a list of one or more test cases
to exercise.

Limited the test-fix-wrong-dns.sh script to work with IPv4 only.

Created ansible task to setup the 'bad_dns' test case by changing
the DNS IP address of master-0 to a different subnet.

Moved the test execution into its own function.

* Replace builtin command with template conditionals

* Fix typo