AGENT-616: Create unconfigured ignition and config-image and mount to VM (#1533)
This provides a test for both https://issues.redhat.com/browse/AGENT-558 and
https://issues.redhat.com/browse/AGENT-559. It runs the installer commands
to create an unconfigured ignition and to create a config-image. It
creates an ISO from the unconfigured ignition suitable for booting the
VM. It also attaches the config-image to the VM which will be loaded
when detected during system boot.