Fix CVO overrides (#1537)
Prior to OCP 4.10, the 'group' field in the override was ignored.
Although "apps/v1" was the documented thing to pass, after
openshift/cluster-version-operator#689 stopped ignoring the field, only
"apps" works.

My proposed fix for https://issues.redhat.com/browse/OCPBUGS-15390 is to
make both work, so we might as well switch to the one that works with
older releases.