Agent: Remove bogus override of EXTERNAL_SUBNET_V6 (#1554)
* Agent: Use correct subnet masks for static IPs

The CLUSTER_HOST_PREFIX_* variables correspond to the clusterNetwork in
the install-config, whereas the IP addresses of the nodes are on the
EXTERNAL_SUBNET_* networks. Since these networks may not have the same
subnet mask, we need to treat them independently.

* Agent: Remove bogus override of EXTERNAL_SUBNET_V6

Not all scripts in a agent-based install source agent/common.sh -
specifically, the configure step 02_configure_host.sh does not, and this
is the step that sets up the DHCP config.

The result of overriding EXTERNAL_SUBNET_V6 in agent/common.sh then is
that we get inconsistent values in different parts of dev-scripts. The
DHCP server is set up with the default value from network.sh, while the
install-config gets the overridden value.

The two values differ only in their prefix length (120 vs. 64).
Unfortunately, assisted refuses to accept that hosts are in the machine
network if the prefix length of the network they are in does not match
exactly the prefix length of the machine network specified in the
install-config. When this happens, assisted rejects the hosts with the
cryptic error "No connectivity to the majority of hosts in the cluster."
(see https://issues.redhat.com/browse/OCPBUGS-7928).

As a result, testing was broken on multinode DHCP scenarios with IPv6,
including the standard HA_IPV4V6_DHCP scenario we use for
e2e-agent-ha-dualstack in many repositories.

The original reason for this was that static IPs were created using the
subnet mask from the clusterNetwork, instead of the machineNetwork where
they actually appear. With this conflation eliminated, there is no
longer a need to force the two networks to have the same subnet mask.