OCPBUGS-16912: Agent use nmstateconfig for DHCP tests (#1557)
* OCPBUGS-16912: Agent use nmstateconfig for DHCP tests

The Agent IPV6_DCHCP test fails due to the address assigned via DHCPv6
not matching the rendezvousIP. The reason is that by default
NetworkManager sends DHCPv6 Solicitations with a Client ID of DUID-UID
(prefix=00:04, see https://datatracker.ietf.org/doc/html/rfc6355)
containing a machine specific value. When using nmstateconfig for DHCP,
the NetworkManager keyfiles are setup for each interface with dhcp-duid=ll
(prefix=00:03, see https://datatracker.ietf.org/doc/html/rfc3315#section-9.4)
which means that the DHCPv6 Solicitation is based on the mac address. This
allows the DHCPv6 server to assign a deterministic address which
can be set to match the rendezvousIP, e.g
https://github.com/metal3-io/metal3-dev-env/blob/main/vm-setup/roles/libvirt/templates/network.xml.j2#L93

Using the nmstateconfig is also recommended for IPv4 and is consistent with
configurations we have seen from customers. For both cases the nmstate
configs are used - https://nmstate.io/examples.html#dynamic-ip-configuration.

https://github.com/openshift-metal3/dev-scripts/pull/1555 is also needed
and included here.

* Use external_subnet_vX_prefixlen for nmstateconfig

* add new variable for dhcp and use current behaviour by default