AGENT-698: Agent support for OCI platform (#1576)
Initial support for the external OCI platform. This sets the libvirt
domain manufacturer and product to allow the assisted-service
validations for this platform to pass.

Requires https://github.com/openshift/assisted-service/pull/5438
and https://github.com/openshift/installer/pull/7442

In addition, this requires changes to either:
a) use a minimalISO instead of a fullISO in the agent-based-installer
b) not require a mimimalISO in assisted-service