Set net.ipv6.conf.all.accept_ra=2 for all topologies (#1583)
In specific scenarios we deploy dual-stack NICs even if the host is
configured as IPv4 stack (e.g. when testing multi NIC with nmstate). In
this situation we forgot to set accept_ra what causes deployment to
fail.

With this change we will set accept_ra no matter the topology of the
cluster so that we can have all types of VMs created.