OCPBUGS-15056: Use v4 provisioning network for v4-primary dual stack (#1534)
Under some circumstances, using an ipv6 provisioning network in a
v4-primary dual stack env is causing BMO to bring nodes up with
only a v6 address. This is a problem because in v4-primary envs we
only configure the local registry to be accessible on v4.

However, since there's technically no reason you couldn't run this
way if you really wanted, let's also add a v6 DNS entry for the
virthost so the registry is also accessible on v6 in dual stack
clusters.

This is a belt-and-suspenders approach to the fix. Either of these
changes on its own would fix the immediate problem, but doing both
gives us arguably more sane defaults and allows more combinations
of options to work in dev-scripts.