Stop running auxiliary containers with maximum priveleges
None of the Metal3 components actually require --privileged.
Instead, disable the SELinux isolation to allow the containers
to mount host directories.

Also don't expose vbmc to the host networking.

For this all to work, stop using a pod for these containers.
I'm not sure why we even do that in the first place.