METAL-754: Switch testing container image to RHEL9
We also need to maintain RHEL8 support as prow
uses RHEL8 as a base image. Once merged we can
switch the base image (in the release repo) to
rhel9 and remove rhel8 support.

We're moving to RHEL9 now in order to take advantage
of curl's new "--retry-all-errors" flag. This will
allow us to retry when the ofcir-acquire ci step
fails to contact ofcir.

Also remove RHEL7 support, it hasn't been needed
since February.