Move release_info functions to a separate script
Move the functions for creating and interpreting the release_info.json
file from ocp_install_env.sh to a separate script. This simplifies the
dependencies of the scripts that use it. It also allows us to use
this function in other places, where importing ocp_install_env is
currently not required.