CORS-2969: Don't default to OpenShiftSDN after deprecation (#1622)
* Move release_info functions to a separate script

Move the functions for creating and interpreting the release_info.json
file from ocp_install_env.sh to a separate script. This simplifies the
dependencies of the scripts that use it. It also allows us to use
this function in other places, where importing ocp_install_env is
currently not required.

* CORS-2969: Don't default to OpenShiftSDN after deprecation

OpenShiftSDN is being deprecated in 4.15, so switch the default to
OVNKubernetes in the IPv4-only case (OVNk is already required for
IPv6/dual-stack) to avoid installer failures.

This follows up on 8537fdd1e01fa07caa7f09ed03c6beafdacb9378, which sets
OpenShiftSDN in the manifests (but not the install-config) iff it is
*explicitly* requested.

Fixes #1613