AGENT-724: Deploy vSphere clusters with credentials (#1599)
Fake credentials are added to install-config.yaml for vSphere
platform.

In combination with
https://github.com/openshift-metal3/dev-scripts/pull/1588,
this patch allows testng of installer and assisted-service
validations for the vSphere platform.

A complete install for vSphere using dev-scripts is currently
not possible because the vSphere environment is partially
simulated using libvirt VMs. Any cloud controller manager
interactions with vSphere will fail in libvirt because there
isn't an actual vSphere environment that is being exercised.

In this setup, only "agent wait-for bootstrap-complete" will
succeed. For vSphere, "agent wait-for install-complete" will fail
and is skipped.