Extract oc RHEL8 binary from release explicitly (#1646)
In the course of bumping oc's base image to RHEL9, it is needed that
the environments bases on RHEL 8 (at least for now) to use oc RHEL8
binary. This PR enables that when the dev-scripts is also bumped to RHEL9,
it will be an easy change to use oc.rhel9 explicitly.

This PR changes the extraction of oc to oc.rhel8 which exists in 4.15 and
4.16 release payloads and fall backs to default oc for older releases.