CORS-2797: use static openshift-installer when possible
With https://github.com/openshift/installer/pull/8161, the baremetal
installer is now part of the regular installer which is now statically
linked.

Using the static binary has the advantage of working on both CS8 and CS9
VMs and is a necessary step for moving the Installer images to a RHEL9
base.