CORS-3462: extract static openshift-installer from 4.16 (#1653)
* Refactor calculation of default installer cmd for extraction

Dynamically determine the installer command to extract from the payload.
After extraction, rename it to "openshift-install" so that we always
have a consistent command name, regardless of what was extracted or if
we built it locally.

* Switch from openshift-baremetal-install to openshift-install from 4.16

This allows us to continue to run on RHEL/CentOS 8.
openshift-baremetal-install will be dynamically linked for RHEL 9
starting with 4.16.

* Use FIPS-capable installer when requested

Only do this if FIPS_VALIDATE is set to true, otherwise this will break
the ability to test FIPS on RHEL/CentOS 8 with OCP 4.16 and later.