Run subscription-manager command when system is subscribed (#1681)
https://github.com/openshift-metal3/dev-scripts/pull/1674 enables
CRB repos on RHEL using subscription-manager.

RHEL system might not be subscribed to subscription-manager. The system
might be using custom repos to install libvirt-devel package. Then
dev-scripts installation will fail.

In order to fix it. We can use subscription-manager identity command
to check the system status. If a system is subscribed to subscription
manager, then exit status will be 0 else 1. It will help us to use
both the system.

Below is the example for the same:
A system is subscribed with subscription-manager:
```
[zuul@ospci-baremetal-01 ~]$ sudo subscription-manager identity
system identity: xxx
name: xxx
org name: xxx
org ID: xxx
[zuul@ospci-baremetal-01 ~]$ echo $?
0
```
A system with no subscrition:
```
[root@ospci-baremetal-02 ~]# subscription-manager identity
This system is not yet registered. Try 'subscription-manager register --help' for more information.
[root@ospci-baremetal-02 ~]# echo $?
1
```

Signed-off-by: Chandan Kumar <raukadah@gmail.com>