Don't force use of a mirrored image in v6-primary (#1684)
It isn't necessary to use a mirror with v6-primary since it's
still dual stack and still has access to v4 public registries.
By forcing the use of the mirrored image path, we actually break
the no mirror use case because we point the deployment at an image
that doesn't exist in the local registry.