Fix mirror config for v6-primary
It turns out we do need to mirror images for v6-primary clusters
because during provisioning they only get an IP from the primary
family, which means they can't pull images from the public ipv4
image registries. This fixes the logic so it is all consistent
where it needs to be so v6-primary deployments will work as intended
by default.