Don't restart NetworkManager unnecessarily (#1762)
Restarting NM breaks VPN connections, which are required to build
images for some OpenShift projects, notably machine-config-operator.
Forcing a NetworkManager restart every time we update the node, even
when no changes to the NetworkManager package were made, causes
problems when working on patches to MCO.

This change checks the NM version before and after the dnf update
and only restarts NM if the version changed.