OCPBUGS-55386: Update oc-mirror to use -v2 (#1752)
Use version v2 of oc-mirror. This version has many improvements over v1
but the primary reason for using v2 is that it allows nightly or CI
builds to be tested via the release field in the ImageSetConfiguration.

Note that oc-mirror is not currently used in any CI tests and has strictly
been used for local testing with the agent-based installer. As a
result its not necessary to keep backwards compatibility with v1.

It's expected that oc-mirror will eventually be used for setting up a
mirror for disconnected testing instead of 'oc adm release mirror'
since its the most common way to set up mirrors by our customers;
'oc adm release mirror' does not provide for installing additional
operators.