Disable cert validation if needed (#1757)
With the updated version of metal3-dev-env, the redfish endpoint will be TLS secured,
with a self signed certificate. This will be indicated by the https protocol in the
redfish address, and a redfish_verify_ca field set to False. We need to configure BMHs
with disableCertificateVerification = true in this case.

Signed-off-by: Marc Sluiter <msluiter@redhat.com>