Add 3s sleep after ESC from network tree view
The bad_dns tests appears to fail when the Tab key press to
move to the <Quit> button is not being registered. At the end
of the test Enter key is pressed but instead of quiting
agent-tui, it enters nmtui because the cursor is still on the
<Configure Network> button.

Adding a 3s sleep after escaping from the the network tree view
should give it more time to properly exit and return to
agent-tui.