Add an additional ESC key press after exiting nmtui
There appears to be a blank screen in some environments that
that appears after exiting nmtui. The ESC key then exits out of
this blank screen and into the network tree view. Adding
an extra ESC key press to fully exit out of nmtui and tree view.

In environments where the blank screen does not appear, the
extra ESC key press is innocuous.