NO-ISSUE: Take screenshot after each console key press (#1768)
* Take screenshot after each console key press

Add additional screenshots to help with debugging issues with
console interaction in CI tests.

* Add 3s sleep after ESC from network tree view

The bad_dns tests appears to fail when the Tab key press to
move to the <Quit> button is not being registered. At the end
of the test Enter key is pressed but instead of quiting
agent-tui, it enters nmtui because the cursor is still on the
<Configure Network> button.

Adding a 3s sleep after escaping from the the network tree view
should give it more time to properly exit and return to
agent-tui.

* Add an additional ESC key press after exiting nmtui

There appears to be a blank screen in some environments that
that appears after exiting nmtui. The ESC key then exits out of
this blank screen and into the network tree view. Adding
an extra ESC key press to fully exit out of nmtui and tree view.

In environments where the blank screen does not appear, the
extra ESC key press is innocuous.

* Quit agent-tui using the q key

Navigating to the <Quit> button using TAB or LEFT sometimes does
not register when sent through "virsh send-key" in the CI.

https://github.com/openshift/agent-installer-utils/pull/107
introduced an alternative way to exit agent-tui from the checks
page.